style

+++ backend/src/controllers/imagesControllers.ts
@@ -4,13 +4,13 @@ import { randomUUID } from 'crypto';
 import bcrypt from 'bcryptjs';
 import fs from 'fs';
 import { promisify } from 'util';
-// import sharp from 'sharp';
 import cloudinary from '../config/cloudinary';
 
 import { ImageType, ResImage } from '../interfaces';
 
 const unlinkAsync = promisify(fs.unlink);
 
+// getAllImages
 export const getAllImages = async (
   req: Request,
   res: Response,
@@ -25,6 +25,7 @@ export const getAllImages = async (
   return res.status(200).json(images.map((image) => setResImage(image)));
 };
 
+// getSingleImage
 export const getSingleImage = async (
   req: Request,
   res: Response,
@@ -38,13 +39,14 @@ export const getSingleImage = async (
   return res.status(200).json({ image: setResImage(image) });
 };
 
+// createImages
 export const createImages = async (
   req: Request,
   res: Response,
 ): Promise<Response<unknown, Record<string, unknown>>> => {
   const files = req.files as Express.Multer.File[];
   console.log(files);
-  const { label, password } = req.body as { label: string; password: string | undefined };
+  const { label, password, ownerId } = req.body as { label: string; password: string | undefined; ownerId: string };
 
   if (!files || !label) {
     return res.status(400).json({ error: 'missing required property' });
@@ -65,6 +67,7 @@ export const createImages = async (
         password: hash,
         url: uploadImage.public_id,
         isProtected: hash ? true : false,
+        ownerId: ownerId,
       })) as ImageType;
 
       await unlinkAsync(file.path);
@@ -76,10 +79,23 @@ export const createImages = async (
   return res.status(201).json();
 };
 
+// getMyImages
+export const getMyImages = async (req: Request, res: Response): Promise<Response> => {
+  const userId = res.locals.userId as string;
+
+  const myImages: ImageType[] | null = await Image.find({ ownerId: userId });
+  return res.status(200).json({ images: myImages });
+};
+
+// deleteImages
 export const deleteImage = async (req: Request, res: Response): Promise<Response<unknown, Record<string, unknown>>> => {
   const { id } = req.params as { id: string };
-  const { password } = req.body as { password: string | undefined };
+  const { password, ownerId } = req.body as { password: string | undefined; ownerId: string | undefined };
   const image: ImageType | null = await Image.findById(id);
+
+  if (!image) {
+    return res.status(404).json({ error: 'image not found' });
+  }
 
   const isReqCorrect = async (password: string | undefined, hash: string | undefined): Promise<boolean> => {
     if ((hash && !password) || (!hash && password)) {
@@ -97,8 +113,11 @@ export const deleteImage = async (req: Request, res: Response): Promise<Response
     return res.status(401).json({ error: 'invalid credital' });
   }
 
+  if (image.ownerId !== ownerId) {
+    return res.status(401).json({ error: 'user not the owner' });
+  }
+
   const deletedImage = (await Image.findByIdAndDelete(id)) as ImageType;
-  // console.log(deleteImage);
   await cloudinary.uploader.destroy(deletedImage.url, function (error, result) {
     console.log(result, error);
   });
@@ -111,5 +130,6 @@ const setResImage = (image: ImageType): ResImage => {
     url: image.url,
     isProtected: image.isProtected,
     _id: image._id,
+    ownerId: image.ownerId,
   };
 };
